requirements.txt
flask
transformers==4.56
torch==2.8
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.3.2
indic-transliteration==2.3.75
pypinyin==0.55.0
jieba==0.42.1
pythainlp==5.1.2
transliterate==1.10.2
sounddevice==0.5.2
openai-whisper
pyinstaller>=6.11
setuptools<81
